LayoutTests/imported/w3c/web-platform-tests/eventsource/request-credentials-expected.txt

Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=1_1622045234749
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=1_1622045234749 due to access control checks.
Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=2_1622045234749
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=2_1622045234749 due to access control checks.
Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=3_1622045234749
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=3_1622045234749 due to access control checks.

Harness Error (TIMEOUT), message = null

TIMEOUT EventSource: credentials: credentials enabled Test timed out
TIMEOUT EventSource: credentials: credentials disabled Test timed out
TIMEOUT EventSource: credentials: credentials default Test timed out
